split the settings form into a separate comp
